+++ client/src/common/components/Page404.scss
@@ -1,8 +1,4 @@
 .page-not-found {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-
   &__wrapper {
     width: 100%;
     position: absolute;
